Move changes internally to avoid minor version bump
